Implement project_client.get_openai_client(). Add tests for AOAI connection with Entra ID auth
