+++ sdk/ai/azure-ai-projects/azure/ai/projects/_patch.py
@@ -7,13 +7,24 @@
 Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python/customize
 """
 import os
-from typing import List, Any, Optional
+import logging
+from typing import List, Any, Optional, TYPE_CHECKING
 from typing_extensions import Self
+from urllib.parse import urlparse
+from azure.core.tracing.decorator import distributed_trace
 from azure.core.credentials import TokenCredential
 from azure.ai.agents import AgentsClient
 from ._client import AIProjectClient as AIProjectClientGenerated
-from .operations import TelemetryOperations, InferenceOperations
+from .operations import TelemetryOperations
 from ._patch_telemetry import enable_telemetry
+from .models._enums import ConnectionType
+from .models._models import ApiKeyCredentials, EntraIDCredentials
+
+if TYPE_CHECKING:
+    # pylint: disable=unused-import,ungrouped-imports
+    from openai import OpenAI
+
+logger = logging.getLogger(__name__)
 
 _console_logging_enabled: bool = os.environ.get("ENABLE_AZURE_AI_PROJECTS_CONSOLE_LOGGING", "False").lower() in (
     "true",
@@ -47,6 +58,26 @@ def _patch_user_agent(user_agent: Optional[str]) -> str:
     return patched_user_agent
 
 
+def _get_aoai_inference_url(input_url: str) -> str:
+    """
+    Converts an input URL in the format:
+    https://<host-name>/<some-path>
+    to:
+    https://<host-name>
+
+    :param input_url: The input endpoint URL used to construct AIProjectClient.
+    :type input_url: str
+
+    :return: The endpoint URL required to construct an AzureOpenAI client from the `openai` package.
+    :rtype: str
+    """
+    parsed = urlparse(input_url)
+    if parsed.scheme != "https" or not parsed.netloc:
+        raise ValueError("Invalid endpoint URL format. Must be an https URL with a host.")
+    new_url = f"https://{parsed.netloc}"
+    return new_url
+
+
 class AIProjectClient(AIProjectClientGenerated):  # pylint: disable=too-many-instance-attributes
     """AIProjectClient.
 
@@ -54,8 +85,6 @@ class AIProjectClient(AIProjectClientGenerated):  # pylint: disable=too-many-ins
     :vartype agents: azure.ai.agents.AgentsClient
     :ivar connections: ConnectionsOperations operations
     :vartype connections: azure.ai.projects.operations.ConnectionsOperations
-    :ivar inference: InferenceOperations operations
-    :vartype inference: azure.ai.projects.operations.InferenceOperations
     :ivar telemetry: TelemetryOperations operations
     :vartype telemetry: azure.ai.projects.operations.TelemetryOperations
     :ivar datasets: DatasetsOperations operations
@@ -88,7 +117,6 @@ class AIProjectClient(AIProjectClientGenerated):  # pylint: disable=too-many-ins
         super().__init__(endpoint=endpoint, credential=credential, **kwargs)
 
         self.telemetry = TelemetryOperations(self)  # type: ignore
-        self.inference = InferenceOperations(self)  # type: ignore
         self._agents: Optional[AgentsClient] = None
 
     @property
@@ -107,6 +135,127 @@ class AIProjectClient(AIProjectClientGenerated):  # pylint: disable=too-many-ins
                 **self._kwargs,
             )
         return self._agents
+
+    @distributed_trace
+    def get_openai_client(
+        self, *, api_version: Optional[str] = None, connection_name: Optional[str] = None, **kwargs
+    ) -> "OpenAI":  # type: ignore[name-defined]
+        """Get an authenticated OpenAI client (from the `openai` package) to use with
+        AI models deployed to your AI Foundry Project or connected Azure OpenAI services.
+
+        .. note:: The package `openai` must be installed prior to calling this method.
+
+        :keyword api_version: The Azure OpenAI api-version to use when creating the client. Optional.
+         See "Data plane - Inference" row in the table at
+         https://learn.microsoft.com/azure/ai-services/openai/reference#api-specs. If this keyword
+         is not specified, you must set the environment variable `OPENAI_API_VERSION` instead.
+        :paramtype api_version: Optional[str]
+        :keyword connection_name: Optional. If specified, the connection named here must be of type Azure OpenAI.
+         The returned OpenAI client will use the inference URL specified by the connected Azure OpenAI
+         service, and can be used with AI models deployed to that service. If not specified, the returned
+         OpenAI client will use the inference URL of the parent AI Services resource, and can be used
+         with AI models deployed directly to your AI Foundry project.
+        :paramtype connection_name: Optional[str]
+
+        :return: An authenticated OpenAI client
+        :rtype: ~openai.OpenAI
+
+        :raises ~azure.core.exceptions.ResourceNotFoundError: if an Azure OpenAI connection
+         does not exist.
+        :raises ~azure.core.exceptions.ModuleNotFoundError: if the `openai` package
+         is not installed.
+        :raises ValueError: if the connection name is an empty string.
+        :raises ~azure.core.exceptions.HttpResponseError:
+        """
+        if connection_name is not None and not connection_name:
+            raise ValueError("Connection name cannot be empty")
+
+        try:
+            from openai import AzureOpenAI
+        except ModuleNotFoundError as e:
+            raise ModuleNotFoundError(
+                "OpenAI SDK is not installed. Please install it using 'pip install openai'"
+            ) from e
+
+        if connection_name:
+            connection = self.connections._get_with_credentials(  # pylint: disable=protected-access
+                name=connection_name, **kwargs
+            )
+            if connection.type != ConnectionType.AZURE_OPEN_AI:
+                raise ValueError(f"Connection `{connection_name}` is not of type Azure OpenAI.")
+
+            azure_endpoint = connection.target[:-1] if connection.target.endswith("/") else connection.target
+
+            if isinstance(connection.credentials, ApiKeyCredentials):
+
+                logger.debug(
+                    "[get_openai_client] Creating OpenAI client using API key authentication, on connection `%s`, endpoint `%s`, api_version `%s`",
+                    connection_name,
+                    azure_endpoint,
+                    api_version,
+                )
+                api_key = connection.credentials.api_key
+                client = AzureOpenAI(api_key=api_key, azure_endpoint=azure_endpoint, api_version=api_version)
+
+            elif isinstance(connection.credentials, EntraIDCredentials):
+
+                logger.debug(
+                    "[get_openai_client] Creating OpenAI using Entra ID authentication, on connection `%s`, endpoint `%s`, api_version `%s`",
+                    connection_name,
+                    azure_endpoint,
+                    api_version,
+                )
+
+                try:
+                    from azure.identity import get_bearer_token_provider
+                except ModuleNotFoundError as e:
+                    raise ModuleNotFoundError(
+                        "azure.identity package not installed. Please install it using 'pip install azure.identity'"
+                    ) from e
+
+                client = AzureOpenAI(
+                    # See https://learn.microsoft.com/python/api/azure-identity/azure.identity?view=azure-python#azure-identity-get-bearer-token-provider # pylint: disable=line-too-long
+                    azure_ad_token_provider=get_bearer_token_provider(
+                        self._config.credential,  # pylint: disable=protected-access
+                        "https://cognitiveservices.azure.com/.default",  # pylint: disable=protected-access
+                    ),
+                    azure_endpoint=azure_endpoint,
+                    api_version=api_version,
+                )
+
+            else:
+                raise ValueError("Unsupported authentication type {connection.type}")
+
+            return client
+
+        try:
+            from azure.identity import get_bearer_token_provider
+        except ModuleNotFoundError as e:
+            raise ModuleNotFoundError(
+                "azure.identity package not installed. Please install it using 'pip install azure.identity'"
+            ) from e
+
+        azure_endpoint = _get_aoai_inference_url(
+            self._config.endpoint  # pylint: disable=protected-access
+        )
+
+        logger.debug(
+            "[get_openai_client] Creating OpenAI client using Entra ID authentication, on parent AI Services resource, endpoint `%s`, api_version `%s`",
+            azure_endpoint,
+            api_version,
+        )
+
+        client = AzureOpenAI(
+            # See https://learn.microsoft.com/python/api/azure-identity/azure.identity?view=azure-python#azure-identity-get-bearer-token-provider # pylint: disable=line-too-long
+            azure_ad_token_provider=get_bearer_token_provider(
+                self._config.credential,  # pylint: disable=protected-access
+                "https://cognitiveservices.azure.com/.default",  # pylint: disable=protected-access
+            ),
+            azure_endpoint=azure_endpoint,
+            api_version=api_version,
+        )
+
+        return client
 
     def close(self) -> None:
         if self._agents:
