+++ sdk/ai/azure-ai-projects/azure/ai/projects/operations/_patch.py
@@ -9,12 +9,10 @@ Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python
 """
 from typing import List
 from ._patch_datasets import DatasetsOperations
-from ._patch_inference import InferenceOperations
 from ._patch_telemetry import TelemetryOperations
 from ._patch_connections import ConnectionsOperations
 
 __all__: List[str] = [
-    "InferenceOperations",
     "TelemetryOperations",
     "DatasetsOperations",
     "ConnectionsOperations",
