+++ sdk/ai/azure-ai-projects/samples/inference/async_samples/sample_chat_completions_with_azure_openai_client_async.py
@@ -47,7 +47,7 @@ async def main():
             print(
                 "Get an authenticated Azure OpenAI client for the parent AI Services resource, and perform a chat completion operation:"
             )
-            async with await project_client.inference.get_azure_openai_client(api_version="2024-10-21") as client:
+            async with await project_client.get_openai_client(api_version="2024-10-21") as client:
 
                 response = await client.chat.completions.create(
                     model=model_deployment_name,
@@ -64,7 +64,7 @@ async def main():
             print(
                 "Get an authenticated Azure OpenAI client for a connected Azure OpenAI service, and perform a chat completion operation:"
             )
-            async with await project_client.inference.get_azure_openai_client(
+            async with await project_client.get_openai_client(
                 api_version="2024-10-21", connection_name=connection_name
             ) as client:
