+++ sdk/ai/azure-ai-projects/samples/inference/sample_chat_completions_with_azure_openai_client.py
@@ -43,7 +43,7 @@ with DefaultAzureCredential(exclude_interactive_browser_credential=False) as cre
         print(
             "Get an authenticated Azure OpenAI client for the parent AI Services resource, and perform a chat completion operation:"
         )
-        with project_client.inference.get_azure_openai_client(api_version="2024-10-21") as client:
+        with project_client.get_openai_client(api_version="2024-10-21") as client:
 
             response = client.chat.completions.create(
                 model=model_deployment_name,
@@ -60,7 +60,7 @@ with DefaultAzureCredential(exclude_interactive_browser_credential=False) as cre
         print(
             "Get an authenticated Azure OpenAI client for a connected Azure OpenAI service, and perform a chat completion operation:"
         )
-        with project_client.inference.get_azure_openai_client(
+        with project_client.get_openai_client(
             api_version="2024-10-21", connection_name=connection_name
         ) as client:
