+++ sdk/ai/azure-ai-projects/samples/inference/sample_chat_completions_with_azure_openai_client_and_azure_monitor_tracing.py
@@ -54,7 +54,7 @@ with DefaultAzureCredential(exclude_interactive_browser_credential=False) as cre
 
         configure_azure_monitor(connection_string=application_insights_connection_string)
 
-        with project_client.inference.get_azure_openai_client(api_version="2024-10-21") as client:
+        with project_client.get_openai_client(api_version="2024-10-21") as client:
 
             response = client.chat.completions.create(
                 model=model_deployment_name,
